+++ CHANGELOG.rst
@@ -5,6 +5,11 @@ Changelog
 current (2022-XX-XX)
 --------------------
 
+0.1.1 (2022-11-29)
+--------------------
+* update parser for v15 to handle former v13 key names, also update outdated License (data-)class in oem_v15 structure. (PR#77)
+* change the validation to return a report and enable report file creation option to the arguments of validation method. (PR#81)
+
 0.1.0 (2022-11-18)
 --------------------
 * Add validation and helper functionality - validation based on json schema and the oemetadata schema files that are published for each release (PR#63)